requirements.txt
asgiref==3.11.1
click==8.3.1
croniter==6.0.0
Django==6.0.2
django-filter==25.2
django-rq==3.2.2
djangorestframework==3.16.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.2.0
rq==2.6.1
six==1.17.0
sqlparse==0.5.5
